added Registations to EventManager
Cars with the EventBehavior Script will be Registered to the TrafficEventmanager directly when they are instantiated.

The Eventmanager uses a Singleton Pattern.

Currently a Delegate system seems not to be necessary.

I also added some visual triggers to the scene.

I fixed a rider bug which seem to make problems