fixing (again) wheels
added for safety reasons a prefab for max HUD jeep.
fixed missing wheels again.
added a bezier path just for the purpose of testing the Event Functionality. Its not aligned perfectly!
Make a small Event scenario.

I fixed, because Rider do something strangely and showed me a bug , a small symbol in a script of the Assetbundler. I dont know what , why , it showed an error in c# but not in the unity edtior. If something strange is happening I can revert that ( it was just a line space between a symbol )