Refunctioning ControlSwitch
the Control Switch is now in charge of the ManualControl Scrtipt, if it is added then the Manual Control is only Working if it is enabled.

I am not completely proud of this solution, but right now this is working for me.

Added also a test color for the AI jeep, just to see a difference between true AI and Manual Switch Jeeps.

worked also on the Event Behavior, Management and Trigger Mechanics. I need some restructuring