first iteration of the event system
the participant car is now activating the trigger, the event manager is managing its behavior towards the control switch and is intended to regulate the traffic. Still a lot of the functions need some refinement, we need to rewatch the Get to the closest point functionality. also the needed Trigger restarting the scene are missing.