adjusted the car physics
the changes I did in my testing showed the best results with this configuration:

the brake torque is now 1000 for each wheel (all wheel brake). Brakes are now overall stronger. potentially need adjustments for a realistic braking behavior (bremsweg berechnung)

the acceleration 1500 instead of 1000. More realistic acceleration of the car. Especially in the Autobahn scene

sideway friction of the wheel is 1.2 instead of 1. They were set a bit higher. They reduce the shaking of the car on the wheel and give it a bit more stability. they are still a lot lower than 2.0 which make them feel like tank chains instead of car wheels.